Admin backend group done

+++ src/main/java/com/revature/controllers/ProductController.java
@@ -44,13 +44,25 @@ public class ProductController {
 
     @Authorized
     @PutMapping
-    public ResponseEntity<Product> upsert(@RequestBody Product product, HttpSession session) {
+    public ResponseEntity<String> upsert(@RequestBody Product product, HttpSession session) {
         User u= (User) session.getAttribute("user");
         if(u.getRole().toString()=="Admin") {
-            return ResponseEntity.ok(productService.save(product));
+            productService.save(product);
+            return ResponseEntity.ok("The product is added");
         }
         else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.ok("Must be logged in as Admin to perform this action");
+        }
+    }
+    @Authorized
+    @PostMapping
+    public ResponseEntity<String> update(@RequestBody Product product, HttpSession session) {
+        User u= (User) session.getAttribute("user");
+        if(u.getRole().toString()=="Admin") {
+            return ResponseEntity.ok(productService.update(product));
+        }
+        else{
+            return ResponseEntity.ok("Must be logged in as Admin to perform this action");
         }
     }
 
@@ -83,21 +95,60 @@ public class ProductController {
 
     @Authorized
     @DeleteMapping("/{id}")
-    public ResponseEntity<Product> deleteProduct(@PathVariable("id") int id) {
+    public ResponseEntity<String> deleteProduct(@PathVariable("id") int id, HttpSession session) {
         Optional<Product> optional = productService.findById(id);
-
+        User u= (User) session.getAttribute("user");
         if(!optional.isPresent()) {
             return ResponseEntity.notFound().build();
         }
+        if(u.getRole().toString()=="Admin") {
         productService.delete(id);
-
-        return ResponseEntity.ok(optional.get());
+        }
+        else{
+            return ResponseEntity.ok("Must be logged in as Admin to perform this action");
+        }
+        return ResponseEntity.ok("Product is deleted");
+    }
+    @Authorized
+    @PutMapping("/featured/{id}")
+    public ResponseEntity<String> addFeaturedProduct(@PathVariable("id") int id, HttpSession session) {
+        User u= (User) session.getAttribute("user");
+        if(u.getRole().toString()=="Admin") {
+            return ResponseEntity.ok(productService.addProductToFeaturedList(id));
+        }
+        else{
+            return ResponseEntity.ok("Must be logged in as Admin to perform this action");
+        }
+    }
+    @Authorized
+    @DeleteMapping("/featured/{id}")
+    public ResponseEntity<String> deleteFeaturedProduct(@PathVariable("id") int id, HttpSession session) {
+        User u= (User) session.getAttribute("user");
+        if(u.getRole().toString()=="Admin") {
+            return ResponseEntity.ok(productService.deleteProductToFeaturedList(id));
+        }
+        else{
+            //return ResponseEntity.notFound().build();
+            return ResponseEntity.ok("Must be logged in as Admin to perform this action");
+        }
+    }
+    @Authorized
+    @GetMapping("/featured")
+    public ResponseEntity<List<Product>> getFeaturedProducts() {
+        return ResponseEntity.ok(productService.getFeaturedProducts());
     }
 
     @Authorized
-    @GetMapping("/featured")
-    public ResponseEntity<List<Product>> getFeaturedProducts(@RequestBody boolean f) {
-        return ResponseEntity.ok(productService.getFeaturedProducts(f));
+    @PostMapping("/sale")
+    public ResponseEntity<String> updateSale(@RequestBody Product product, HttpSession session) {
+        User u= (User) session.getAttribute("user");
+        if(u.getRole().toString()=="Admin") {
+            productService.updateSale(product);
+        }
+        else {
+            return ResponseEntity.ok("Must be logged in as Admin to perform this action");
+        }
+        return ResponseEntity.ok("The sale is updated");
     }
 
     @Authorized
